Add support for OCaml 4.09 and 4.10
The .expected files don't have colors anymore, as expected.  It seems
that the `-no-color` option in the test Dune file didn't have the
expected effect, so I replaced it with `OCAML_COLOR=never`.

Since error messages change slightly in 4.09, I had to add a new class
of .expected files for 4.09+.  The expected files for 4.08 are suffixed
with `.408.expected`, the same way as for 4.07.